Use queryset sequence id instead of content id
This commit contains a partial fix restoring compatibility
with querysetsequence==0.9 by dropping use of a private api
in favour of a public one.

Instead of using the content-type id and object-id as the value
for select options, instead the queryset index and object id are
used. The `QuerySetSequence` maintains an ordered list of the
`QuerySet` objects it is composed of, and the index can be used
retrospectively to retrieve the model given the index and pk.

See also: https://github.com/percipient/django-querysetsequence/issues/21#issuecomment-432649275
Partial fix for #1050